+++ src/lib/builders/collapsible/create.ts
@@ -32,10 +32,11 @@ export function createCollapsible(props?: CreateCollapsibleProps) {
 
 	const root = makeElement(name(), {
 		stores: [open, disabled],
-		returned: ([$open, $disabled]) => ({
-			'data-state': $open ? 'open' : 'closed',
-			'data-disabled': disabledAttr($disabled),
-		}),
+		returned: ([$open, $disabled]) =>
+			({
+				'data-state': $open ? 'open' : 'closed',
+				'data-disabled': disabledAttr($disabled),
+			} as const),
 	});
 
 	const trigger = makeElement(name('trigger'), {
@@ -66,12 +67,13 @@ export function createCollapsible(props?: CreateCollapsibleProps) {
 
 	const content = makeElement(name('content'), {
 		stores: [isVisible, open, disabled],
-		returned: ([$isVisible, $open, $disabled]) => ({
-			'data-state': $open ? 'open' : 'closed',
-			'data-disabled': disabledAttr($disabled),
-			hidden: $isVisible ? undefined : true,
-			style: $isVisible ? undefined : styleToString({ display: 'none' }),
-		}),
+		returned: ([$isVisible, $open, $disabled]) =>
+			({
+				'data-state': $open ? 'open' : 'closed',
+				'data-disabled': disabledAttr($disabled),
+				hidden: $isVisible ? undefined : true,
+				style: $isVisible ? undefined : styleToString({ display: 'none' }),
+			} as const),
 	});
 
 	return {