+++ src/tests/calendar/Calendar.spec.ts
@@ -720,7 +720,7 @@ describe('Calendar', () => {
 		}
 	});
 
-	test('dynamically change weekStartsOn option', async () => {
+	test('dynamically change fixedWeeks option', async () => {
 		const { getByTestId, queryByTestId, user } = setup();
 
 		const nextButton = getByTestId('next-button');